extracted logic from statisticsscreen and run tests
